requirements.txt
python-telegram-bot
boto3
python-dotenv